+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/screens/account/info/AccountInfoUiState.kt
@@ -1,5 +1,8 @@
 package net.nymtech.nymvpn.ui.screens.account.info
 
+import net.nymtech.nymvpn.ui.screens.account.info.components.BandwidthUiState
+import net.nymtech.nymvpn.ui.screens.settings.components.SubscriptionUiState
+
 data class AccountInfoUiState(
 	val isLoading: Boolean = true,
 	val isMnemonicStored: Boolean = false,
@@ -9,4 +12,6 @@ data class AccountInfoUiState(
 	val deviceId: String = "",
 	val accountLinkUrl: String? = null,
 	val manageUrl: String? = null,
+	val subscription: SubscriptionUiState? = null,
+	val bandwidth: BandwidthUiState? = null,
 )