+++ nym-vpn-app/src/screens/settings/Settings.tsx
@@ -10,7 +10,6 @@ import { MsIcon, PageAnim, SettingsMenuCard, Switch } from '../../ui';
 import { AccountSettingRow } from './account';
 import { InfoData } from './info-data';
 import SettingsGroup from './SettingsGroup';
-import Logout from './Logout';
 
 function Settings() {
   const {
@@ -203,7 +202,6 @@ function Settings() {
           },
         ]}
       />
-      <Logout />
       <SettingsMenuCard title={t('quit')} onClick={exit} />
       <InfoData />
     </PageAnim>