+++ nym-vpn-app/src/util.ts
@@ -11,6 +11,10 @@ export function capFirst(string: string) {
   return string.charAt(0).toUpperCase() + string.slice(1);
 }
 
+export function formatGb(gb: number | bigint) {
+  return `${Number(gb).toLocaleString()} GB`;
+}
+
 // Given a set of strings, return the strings concatenated by a white space
 export function setToString(obj: Record<string, string>): string {
   return Object.values(obj).reduce((prev, s) => `${prev} ${s}`, '');