+++ nym-vpn-app/CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Ad blocker
 - Dynamic tray icon and menu reflecting current tunnel configuration
+- See bandwidth usage and subscription expiry/renewal date
 - Diagnostic settings
 
 ### Fixed